# api/src/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/admin/test": {"get": {"summary": "C95Cfd3Bc9B04E0Db909551Fcdf612D9", "operationId": "c95cfd3bc9b04e0db909551fcdf612d9_admin_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/admin/dashboard": {"get": {"summary": "B7D8Ecdbb5694126Ba9Cc4Db743Bffa7", "operationId": "b7d8ecdbb5694126ba9cc4db743bffa7_admin_dashboard_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/admin/login": {"post": {"summary": "Cdb9034Dec454E98A5B486213632D623", "operationId": "cdb9034dec454e98a5b486213632d623_admin_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_cdb9034dec454e98a5b486213632d623_admin_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "header_group": "honeypot"}}, "/auth/test": {"get": {"summary": "2Fbb80Bc51994Ec2A8Eb3F2De3Cfa418", "operationId": "2fbb80bc51994ec2a8eb3f2de3cfa418_auth_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/auth/login": {"get": {"summary": "43Ef1277A34E4Df0854E7159733E868D", "operationId": "43ef1277a34e4df0854e7159733e868d_auth_login_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}, "post": {"summary": "Fd6D6511Acd34193A7F373Debeb68199", "operationId": "fd6d6511acd34193a7f373debeb68199_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_fd6d6511acd34193a7f373debeb68199_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "header_group": "honeypot"}}, "/backup/test": {"get": {"summary": "7251E656D99743479657F84806024Cc5", "operationId": "7251e656d99743479657f84806024cc5_backup_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/etc/test": {"get": {"summary": "31E56A89E34D4Acfa2Fdff052F78186A", "operationId": "31e56a89e34d4acfa2fdff052f78186a_etc_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/git/test": {"get": {"summary": "44Ae267A6Eb64570B3Ed824165Fc2F91", "operationId": "44ae267a6eb64570b3ed824165fc2f91_git_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/legacy/test": {"get": {"summary": "6B16Cd4Db8904C4D9A41328A68B8321D", "operationId": "6b16cd4db8904c4d9a41328a68b8321d_legacy_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/old/test": {"get": {"summary": "E2Aa135C0E884A0B882F3Edaf5148C52", "operationId": "e2aa135c0e884a0b882f3edaf5148c52_old_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/protected/test": {"get": {"summary": "71796F1Dced8474Ebd8Ee4F3Afbabc10", "operationId": "71796f1dced8474ebd8ee4f3afbabc10_protected_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/protected/secret": {"get": {"summary": "21B00166A7D8421Ebf0566Fa4E1B0A15", "operationId": "21b00166a7d8421ebf0566fa4e1b0a15_protected_secret_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/protected/secret/unlock": {"post": {"summary": "D8003E1C9Edd4833933067F1Bb41F05A", "operationId": "d8003e1c9edd4833933067f1bb41f05a_protected_secret_unlock_post", "requestBody": {"content": {"application/json": {"schema": {"type": "string", "title": "Password"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "header_group": "honeypot"}}, "/test": {"get": {"summary": "A61Fe69047C04Cd5B4D71D550942B7B6", "operationId": "a61fe69047c04cd5b4d71d550942b7b6_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/.git/config": {"get": {"summary": "3A7Da2423B36455Eb34E9Ba7Acb503Ae", "operationId": "3a7da2423b36455eb34e9ba7acb503ae__git_config_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/etc/passwd": {"get": {"summary": "047B8A33103D42Dba9142E17C0396B17", "operationId": "047b8a33103d42dba9142e17c0396b17_etc_passwd_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/wp-login.php": {"get": {"summary": "D45D0A22707A4Af6A0633942917C3Edb", "operationId": "d45d0a22707a4af6a0633942917c3edb_wp_login_php_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}}, "header_group": "honeypot"}, "post": {"summary": "61C3562Bb9Aa43F2B42A27A4E6Cbc70C", "operationId": "61c3562bb9aa43f2b42a27a4e6cbc70c_wp_login_php_post", "parameters": [{"name": "log", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "Log"}}, {"name": "pwd", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "Pwd"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "header_group": "honeypot"}}, "/phpmyadmin/index.php": {"get": {"summary": "D02E7C459659464C98B10488D9F6F24A", "operationId": "d02e7c459659464c98b10488d9f6f24a_phpmyadmin_index_php_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/backup/db.zip": {"get": {"summary": "4A7203E027494444A89Bb94D4Bf5B63F", "operationId": "4a7203e027494444a89bb94d4bf5b63f_backup_db_zip_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "honeypot"}}, "/exhibit/test": {"get": {"summary": "0C39F689B2E0479386Ce275884D80766", "operationId": "0c39f689b2e0479386ce275884d80766_exhibit_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "exhibit"}}, "/exhibit/feed/test": {"get": {"summary": "69A54274851C46Fb8E7A314A72D03Cc7", "operationId": "69a54274851c46fb8e7a314a72d03cc7_exhibit_feed_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "exhibit"}}, "/exhibit/ip/test": {"get": {"summary": "Cd18Ecd29E724Ae799E1E4516B9895F7", "operationId": "cd18ecd29e724ae799e1e4516b9895f7_exhibit_ip_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}, "header_group": "exhibit"}}, "/exhibit/ip/{ip_addr}/activity": {"get": {"summary": "Fa1Eec2Aeb1744Aa940B03Df39E3448C", "operationId": "fa1eec2aeb1744aa940b03df39e3448c_exhibit_ip__ip_addr__activity_get", "parameters": [{"name": "ip_addr", "in": "path", "required": true, "schema": {"type": "string", "title": "Ip Addr"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "take", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Take"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ListIpActivityRow"}, "title": "Response Fa1Eec2Aeb1744Aa940B03Df39E3448C Exhibit Ip  Ip Addr  Activity Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "header_group": "exhibit"}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}}}}, "components": {"schemas": {"Body_cdb9034dec454e98a5b486213632d623_admin_login_post": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "Body_cdb9034dec454e98a5b486213632d623_admin_login_post"}, "Body_fd6d6511acd34193a7f373debeb68199_auth_login_post": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "Body_fd6d6511acd34193a7f373debeb68199_auth_login_post"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ListIpActivityRow": {"properties": {"event_at": {"type": "string", "format": "date-time", "title": "Event At"}, "event_type": {"type": "string", "title": "Event Type"}, "target_id": {"type": "integer", "title": "Target Id"}, "target_slug": {"type": "string", "title": "Target Slug"}, "details": {"type": "string", "title": "Details"}}, "type": "object", "required": ["event_at", "event_type", "target_id", "target_slug", "details"], "title": "ListIpActivityRow"}, "Message": {"properties": {"detail": {"type": "string", "title": "Detail"}}, "type": "object", "required": ["detail"], "title": "Message"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}